feat：将旧设置基本转移到Setting中，修复了之前遗留的点光源被裁剪的问题，原因为未正确传入每个区块中光源默认数量 并添加了说明注释

+++ Assets/CustomRenderPipeLine/Runtime/RenderGraphPasses/SetUpPass.cs
@@ -46,9 +46,9 @@ public class SetUpPass
 
         //延迟设置Camera缓冲区大小到setup结束
         cmd.SetGlobalVector(_bufferSizeID, new Vector4(
-            1f / _attachmentSize.x, 1f / _attachmentSize.y,
+            1f / (float)_attachmentSize.x, 1f / (float)_attachmentSize.y,
             _attachmentSize.x, _attachmentSize.y));
-
+        
         context.renderContext.ExecuteCommandBuffer(cmd);
         context.cmd.Clear();
     }
